Update change type to int

+++ src/main/java/csd/roster/response_model/SummaryResponseModel.java
@@ -11,17 +11,17 @@ import lombok.*;
 public class SummaryResponseModel {
     // Total number of employees for a certain company
     private int employeesCount;
-    private float employeesCountChange;
+    private int employeesCountChange;
 
     // Total number of employees on leave for a certain company
     private int leaveCount;
-    private float leaveCountChange;
+    private int leaveCountChange;
 
     // Total number of employees working on site
     private int onsiteCount;
-    private float onsiteCountChange;
+    private int onsiteCountChange;
 
     // Total number of employees contracted with covid 
     private int covidCount;
-    private float covidCountChange;
+    private int covidCountChange;
 }
